requirements[1].txt
python-telegram-bot==20.7
selenium==4.19.0